requirements.txt
Babel==2.4.0
Django==1.11.3
django-filter==1.0.4
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
Markdown==2.6.8
phonenumberslite==8.6.0
pytz==2017.2
